Allow rpm tools on RHEL < 8 to parse the spec file
We use some automated tools to parse the spec file, which do not
recognize the newer "Recommends" and "Suggests" tags because they
run on RHEL 7. To allow these tools to work while keeping the
functionality in the spec, let's enclose the tags in an %if clause.

Change-Id: I508ab7679e40259f8f986c8b1f6378024e5bfa2f
(cherry picked from commit e62855db1010118604931f9e8e273fb40bbbb293)